Added new tests and created a new file for Matrix's subclasses
